test(integration): share process stop helper
Move robust child-process teardown into `integration-tests/helpers`
as `stopProc` and use it in debugger test utils.

This centralizes SIGTERM/SIGKILL shutdown handling and helps
prevent cross-test leakage from lingering processes.
